+++ lib/src/ui/ui.dart
@@ -1,0 +1,1 @@
+export 'daily_local_notification_widget.dart';
